feat: Google Classroom integration
This commit introduces Google oAuth2 for Google Classroom course and
assignment linking so students can connect everything together. This
is currently work in progress, but it's definitely a right step in the
right direction.

Currently, it saves all courses (assignments fetched and attached) to
a file, for index.js to read when necessary.